исправляет задание согласно замечаниям
