+++ js/form.js
@@ -2,12 +2,9 @@ import {isEscapeKey, isEnterKey} from './util.js';
 
 /*Константы*/
 
-const imgUploadForm = document.querySelector('.img-upload__form');
 const imgUploadOverlay = document.querySelector('.img-upload__overlay');
 const imgUploadLabelId = document.querySelector('#upload-file');
 const imgUploadCloseButton = document.querySelector('.img-upload__cancel');
-const commentInput = imgUploadForm.querySelector('.text__description');
-const hashtagInput = imgUploadForm.querySelector('.text__hashtags');
 
 /*Открытие окна загрузки*/
 
@@ -56,88 +53,3 @@ function closeUploadOverlay () {
   imgUploadCloseButton.removeEventListener('keydown', onCloseButtonClick);
   imgUploadCloseButton.removeEventListener('keydown', onCloseButtonEnterKeydown);
 }
-
-/*Отмена работы ESС при фокусе (не получилось; долго по-разному пробовала и никак; это последний вариант)*/
-
-/*const onEvtTargetEscKeydown = (evt) => {
-  if (((isEscapeKey(evt)) && (evt.target === hashtagInput)) || ((isEscapeKey(evt)) && (evt.target === commentInput))) {
-    evt.preventDefault();
-    evt.stopPropagation();
-  }
-};
-
-EventTarget.addEventListener('keydown', onEvtTargetEscKeydown);*/
-
-/*Валидация формы*/
-
-const pristine = new Pristine(imgUploadForm, {
-  classTo: 'img-upload__field-wrapper',
-  errorClass: 'img-upload__field-wrapper_invalid',
-  successClass: 'img-upload__field-wrapper_valid',
-  errorTextParent:'img-upload__field-wrapper',
-  errorTextTag: 'p',
-  errorTextClass: 'input__error'
-});
-
-/*Комментарии*/
-
-function matchCommentSymbolsAmount() {
-  return commentInput.value.length <= 140;
-}
-
-pristine.addValidator(
-  imgUploadForm.querySelector('.text__description'),
-  matchCommentSymbolsAmount,
-  'Количество символов в комментарии не должно превышать 140'
-);
-
-/*Хеш-тэги*/
-
-/*Проверка хэш-тегов на кол-во не более 5*/
-
-function matchHastagsAmount() {
-  return hashtagInput.value.split(' ').length <= 5;
-}
-
-pristine.addValidator(
-  imgUploadForm.querySelector('.text__hashtags'),
-  matchHastagsAmount,
-  'Количество хэш-тегов не должно превышать 5'
-);
-
-/*Проверка хэш-тегов на соответствие заданным параметрам:
-- хэш-тег начинается с символа # (решётка);
-- строка после решётки должна состоять из букв и чисел и не может содержать пробелы, спецсимволы (#, @, $ и т. п.),
-символы пунктуации (тире, дефис, запятая и т. п.), эмодзи и т. д.;
-- хеш-тег не может состоять только из одной решётки;
-- максимальная длина одного хэш-тега 20 символов, включая решётку;*/
-
-const regularExpression = /^#[A-Za-zА-Яа-яЁё0-9]{1,19}$/;
-function checkHastagsСontent () {
-  return hashtagInput.value.split(' ').some((hashtag) => regularExpression.test(hashtag));
-}
-
-pristine.addValidator(
-  imgUploadForm.querySelector('.text__hashtags'),
-  checkHastagsСontent,
-  'Хэш-тег должен иметь определённый синтаксис'
-);
-
-/*Проверка хэш-тегов на неповторение*/
-
-function areHashtagsUnique() {
-  return (new Set(hashtagInput.value.split(' '))).size === hashtagInput.value.split(' ').length;
-}
-
-pristine.addValidator(
-  imgUploadForm.querySelector('.text__hashtags'),
-  areHashtagsUnique,
-  'Хэш-теги не должны повторяться'
-);
-
-/*Обработчик на форму*/
-
-imgUploadForm.addEventListener('submit', (evt) => {
-  evt.preventDefault();
-  pristine.validate();
-});
