+++ js/main.js
@@ -2,3 +2,4 @@ import './data.js';
 import './render.js';
 import './render-big-pic.js';
 import './form.js';
+import './hashtags-validation.js';
